beardog_tunnel::tunnel::genetic_healing

Trait EcosystemComputeExtension

pub trait EcosystemComputeExtension: Send + Sync {
    // Required methods
    fn coordinate_network_healing<'life0, 'life1, 'async_trait>(
        &'life0 self,
        issue_type: SecurityIssueType,
        local_healing: &'life1 HealingResult,
    ) -> Pin<Box<dyn Future<Output = BearDogResult<NetworkHealingResult>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn optimize_crypto_algorithms<'life0, 'life1, 'async_trait>(
        &'life0 self,
        performance_metrics: &'life1 PerformanceMetrics,
    ) -> Pin<Box<dyn Future<Output = BearDogResult<CryptoOptimization>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn evolve_network_genetics<'life0, 'async_trait>(
        &'life0 self,
        mutation_trigger: MutationTrigger,
    ) -> Pin<Box<dyn Future<Output = BearDogResult<NetworkGenetics>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Extension trait for universal ecosystem integration This provides the interface for distributed genetic healing across any ecosystem services

Required Methods

fn coordinate_network_healing<'life0, 'life1, 'async_trait>( &'life0 self, issue_type: SecurityIssueType, local_healing: &'life1 HealingResult, ) -> Pin<Box<dyn Future<Output = BearDogResult<NetworkHealingResult>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Network-wide genetic healing coordination

fn optimize_crypto_algorithms<'life0, 'life1, 'async_trait>( &'life0 self, performance_metrics: &'life1 PerformanceMetrics, ) -> Pin<Box<dyn Future<Output = BearDogResult<CryptoOptimization>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Distributed crypto algorithm optimization

fn evolve_network_genetics<'life0, 'async_trait>( &'life0 self, mutation_trigger: MutationTrigger, ) -> Pin<Box<dyn Future<Output = BearDogResult<NetworkGenetics>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Cross-node genetic evolution