beardog_tunnel::tunnel

Module hsm

HSM integration and management

HSM Module - Canonical Redirection

MODERNIZED: REDIRECTS TO UNIVERSAL HSM SYSTEM

This module now serves as a compatibility layer that redirects to the canonical Universal HSM system. All fragmented implementations have been consolidated into the universal system.

Re-exports

pub use crate::universal_hsm::traits::UniversalHsmProvider as HsmProvider;

pub use crate::universal_hsm::UniversalHsmManager as HsmManager;

pub use crate::universal_hsm::providers::software;

pub use crate::universal_hsm::providers::Pkcs11Provider;

pub use crate::universal_hsm::providers::TpmProvider;

pub use crate::universal_hsm::*;

Modules

providersDeprecated

DEPRECATED: Use crate::universal_hsm::providers instead

typesDeprecated

DEPRECATED: Use crate::universal_hsm::traits instead